fix: restructure SmartChat component for clarity
