+++ frontend/components/search-wizard/modes/smart-chat/chat-panel/constants.ts
