+++ frontend/components/search-wizard/modes/smart-chat/chat-panel/index.tsx
@@ -1,24 +1,98 @@
 "use client";
 
+import { useCallback, useEffect, useRef, useState } from "react";
 import { Bot, Loader2 } from "lucide-react";
 
+import { useSearchWizard } from "@contexts/search-wizard";
+import { useIntakeSessions } from "@hooks/use-intake-sessions";
+import { getApiErrorMessage } from "@lib/api-errors";
+import type { LlmInputResponse } from "@services/intake-sessions";
+
+import { INTRO_MESSAGE } from "./constants";
+import type { ChatMessage } from "../types";
 import { ChatComposer } from "./chat-composer";
-import { useSmartChat } from "@contexts/smart-chat";
 import { MessageList } from "./message-list";
 import { styles } from "../styles";
 
-export const ChatPanel = () => {
+type ChatPanelProps = {
+  onLlmSuccess: (data: LlmInputResponse) => void;
+};
+
+export const ChatPanel = ({ onLlmSuccess }: ChatPanelProps) => {
+  const { submitLlmInput } = useIntakeSessions();
   const {
-    busyBootstrap,
-    draft,
     errorMessage,
-    handleSend,
-    isSending,
-    messages,
+    isLoadingQuestion,
+    isSubmitting,
     sessionId,
-    setDraft,
-    textareaRef,
-  } = useSmartChat();
+    setErrorMessage,
+  } = useSearchWizard();
+
+  const [messages, setMessages] = useState<ChatMessage[]>([]);
+  const [draft, setDraft] = useState("");
+  const [isSending, setSending] = useState(false);
+  const textareaRef = useRef<HTMLTextAreaElement>(null);
+  const welcomeSeeded = useRef(false);
+
+  const isBusy = isLoadingQuestion && !sessionId;
+
+  useEffect(() => {
+    if (!sessionId || welcomeSeeded.current) {
+      return;
+    }
+    welcomeSeeded.current = true;
+    setMessages([
+      {
+        id: crypto.randomUUID(),
+        role: "assistant",
+        content: INTRO_MESSAGE,
+      },
+    ]);
+  }, [sessionId]);
+
+  useEffect(() => {
+    textareaRef.current?.focus();
+  }, [sessionId, messages.length]);
+
+  const handleSend = useCallback(async () => {
+    const text = draft.trim();
+    if (!sessionId || !text || isSending || isSubmitting) {
+      return;
+    }
+
+    setSending(true);
+    setErrorMessage(null);
+    setDraft("");
+
+    const userMsg: ChatMessage = {
+      id: crypto.randomUUID(),
+      role: "user",
+      content: text,
+    };
+    setMessages((m) => [...m, userMsg]);
+
+    try {
+      const data = await submitLlmInput(sessionId, {
+        input: text,
+        mode: "llm",
+      });
+      onLlmSuccess(data);
+    } catch (err) {
+      setErrorMessage(getApiErrorMessage(err));
+      setMessages((m) => m.filter((x) => x.id !== userMsg.id));
+      setDraft(text);
+    } finally {
+      setSending(false);
+    }
+  }, [
+    draft,
+    isSending,
+    isSubmitting,
+    onLlmSuccess,
+    sessionId,
+    setErrorMessage,
+    submitLlmInput,
+  ]);
 
   return (
     <div className={styles.chatColumn}>
@@ -33,7 +107,7 @@ export const ChatPanel = () => {
         <div className={`${styles.errorBanner} mx-4 mt-3 sm:mx-5`}>{errorMessage}</div>
       )}
 
-      {busyBootstrap ? (
+      {isBusy ? (
         <div className={styles.loadingWrap}>
           <Loader2 className="mr-2 size-5 animate-spin" aria-hidden />
           Connecting…
@@ -47,7 +121,7 @@ export const ChatPanel = () => {
             textareaRef={textareaRef}
             sessionId={sessionId}
             isSending={isSending}
-            onSend={() => void handleSend()}
+            onSend={handleSend}
           />
         </>
       )}
