+++ frontend/components/search-wizard/modes/smart-chat/side-panel/utils.ts
